Converted all binary expressions into one AST node.
First, apologies for the huge commit. Not a lot of it worked
without everything else. Secondly, justifications:

This is basically a trade off in favour of a smaller AST versus
more code in the backend. This makes things easier for visitors
and the like. As an added bonus, associativity is now correctly
handled and && and || evaluation semantics are handled explicitly
as opposed to working by happy happenstance as before.

To make the code simpler I also converted AssignExpression into
a BinaryExpression also. I think this shouldn't have fucked
anything up too badly, but I could've missed something.

Going forward, UnaryExpression and PostfixExpression could be
handled in the same manner if that is deemed favourable.