Now each declaration is processed in its own fiber.
If a declarationr equire some other processing to continue, it does signal it to the scheduler.
The scheduler will save state and yield, to process the next item in the list.
As of now, scheduler may infinite loop in case of circular dependancies, but
have all data required to not do so. It is expected not to do so anymore in the future.