Add afcility to require from the backend.
The code need to fixup the module to make sure it is in a valid state, but there is no guarantee that the JIT will not run into the unreachable IR added for the fixup. I need to figure out a better wayt o do this, but I'm not sure at this point what it is.