Remove a ton of global state in the backend.
Introduce passes for local and global symbols, instead of one pass for symbol that mixes both. The passes are value types and contains the state required. A reference to the local pass is passed down to StatementGen and ExpressionGen in order for them to have access to the local context which was previously global to the backend.