Add a pass for flow analysis instead of doing it as we process statements.
This effectively make the compiler do 2 passes on statements instead of one, so we should consider having a flat representation for statement in the IR instead of tree like as we have now. This would also create opportunity for a fast interpretor instead of using the JIT 100% of the time.

Having 2 passes on statement open the door to precompute some data instead of discovering it lazilly in the backend, for instance, closure data layout.