Use 2 different backend for JIT and codegen.
This became necessary because there is no way to fixup debug infos in LLVM ATM. This was something needed to be done at some point anyway as CTFE require extra "runtime" checks For instance, pointer are comparable if they are part of the same allocation, not otherwize, meaning the codegen must be different for pointer comparaisons.

The instanciation of the backend and the semantic pass are somewhat improved in the process, but it is clear that the backend needs restructuring.