Make the Name mechanism capable of outputing zero terminated strings.
This ended up engendring massive rework in the lexer, but it is for the best. Hopefully we can get rid of CharPumper. It was here to handle generics input ranges, but it is clear that there is too much optimizations to do on string and genericity come second at this stage.

It also most likely make sense to move mangling into names. They would benefit from instant toStringz as well as reducing the size of Symbol objects.