Remove the hasThis flag from constructors.
The this pointer for constructors is kind of magic and needs to be handled in a different manner anyway. It also causes problem with various behavior of hasThis kicking in and having to be handled specifically.

Bonus points because now, this pointers can be handled with less code in the backend.