Change ABI for large structs construction.
All struct used to be constructed by value, but in practice it only make sense for small structs. For large struct this is horribly innefiscient. As a result a "isSmall" was added on struct and is used to change ABI.

Right now the criterion is that a struct is small if its size is smaller than 256bits, but we probably want to refine this in the future.