Add snap package definition for SDC
This package definition provides a complete recipe for a bootstrapped
build of SDC, with fresh builds of DMD/druntime/phobos 2.072.2 and LLVM
3.9.1.  The resulting SDC and accompanying libraries are then packaged
into a classic snap (meaning its confinement allows it to access the
host system for access to development libraries, toolchains, etc.).

This should provide a straightforward way to test-build SDC as well as
providing an easy way to install the compiler that should be viable for
any Linux distribution with a sufficiently up-to-date snapd package.

Some notes on the parts of the snap package:

  * `llvm` downloads an LLVM source package and performs a full
    build which used as a build dependency by `sdc`, but which
    is not necessary to include in the final snap package.

  * `dmd` builds a fresh copy of DMD 2.072.2, which is then used
    to build `druntime` and `phobos`.  Together these are used
    to build `sdc` itself, but they are not included in the final
    snap package.

  * `snap-config` provides a `dmd.conf` that will be used by the
    bootstrap DMD in order to be able to access the bootstrap
    druntime and phobos, and an `sdc.conf` that specifies how the
    packaged SDC can access its associated libraries and import
    locations.  Only the `sdc.conf` is included in the final snap
    package.

  * `sdc` builds the actual SDC compiler from the HEAD of the
    git repo, using the bootstrap `dmd`, `druntime` and `phobos`,
    and with `llvm` as a build dependency.  The results included
    in the final snap package are specified in the `artifacts:`
    section, with some of them being renamed according to the
    `organize:` config.

The snap `version:` is given as HEAD (yes, the actual string `HEAD`!),
since for now there is no mechanism to extract more informative version
information from git.  Support for this is apparently in the `snapcraft`
roadmap so it should be possible to replace `HEAD` at some point.