Implement proper hexadecimal float decoding.
The ldexp approach had some issues with rounding in some specific cases.
It is still used in cases where there are no rounding problem, but we
fall back to a similar apporach to the one we use with decimal floats
when there is.

Unfortunately, there is a fair bit of duplication between the two, but
also neough difference so that it's not easy to factor. I'm not too
happy with this state of affair, but this will have to do for now.