Have rt.thread scan the stack and static ranges instead of the GC.
The GC request it from rt.thread and pass a callback down. This ensures
we have an API that is compatible with druntime. This is a partial
implementation as static data are not going to be scanned properly, but
it is good enough to wire this into druntime.

We compile libdmdalloc for use with DMD and druntime.