Add API for druntime (#377)
This is a preliminary PR to see how far we are. With this PR, I have a
matching druntime branch which builds against SDC GC.

If we could merge this into master, then we can remove the druntimegc
branch.

The appropriate branch to use for DMD for testing is
https://github.com/schveiguy/dmd/tree/testnewsdcarraystuff

I have no idea how to make tests for this -- it is very dependent on an
unreleased branch of dmd.