[GC] Parse /proc/self/task/<TID>/status to check signal state.
If a thread blocks the signals we use for the GC, then we simply detach
this thread. Detached threads are not stopped by the GC, but we do a
best effort to stack TLS segments for it. Scannign the stack is not
possible as the thread would have to cooperate to indicate its bounds.