Zero slabs that are freed that contain pointers, so when they are allocated next time, they are already zero. (#395)
This isn't as performant as I would like it to be. I can't think of a
good way to run both finalization and zeroing in the same pass.

I also think there is an opportunity to memset batches of slabs, which
might happen if you for instance free an entire graph of nodes at once.
I might be able to get something like that working.