Always terminate helper thread when testing suspend. (#400)
While working on #399, I found that a failed unit test here resulted in
a segfault.

What happens is the assert exits the test, without stopping the helper
thread. Then the stack is now gone, and the helper thread is using
unallocated memory.

Segfaults in sdunit are hard to diagnose because gdb doesn't have any
clue what the call stack means.